The Great Event Rework
The class formerly called Event is now SingularEvent. Event is now
an interface that contains all the methods necessary for the
calendar to display the object. SingularEvent, RecurrentEvent, and
EphemeralEvent all implement the Event interface. Thus, displaying
an event is now decoupled from the implementation. This is done so
that EphemeralEvent can merely point to its parent's attributes
and only copy that information which is unique to it. This should
make all further event work more sensible.